refactor: split cmd/importer/util package
